[📚🎬] MOVIE-APP : React Router 설치 및 레포 정리
